chore: refactor to use a custom hook
